refactor: update entity data fetching and improve TypeScript configuration
