--- web/app/middleware.ts
@@ -1,25 +1,0 @@
-import { NextRequest, NextResponse } from "next/server";
-
-const PUBLIC_PATHS = ["/", "/login", "/api", "/_next", "/favicon.ico"];
-
-export function middleware(request: NextRequest) {
-  const { pathname } = request.nextUrl;
-
-  // Allow public paths and static assets
-  if (PUBLIC_PATHS.some((p) => pathname.startsWith(p))) {
-    return NextResponse.next();
-  }
-
-  // Only protect /app routes
-  if (pathname.startsWith("/app")) {
-    const supabaseToken = request.cookies.get("sb-access-token");
-    if (!supabaseToken) {
-      const url = request.nextUrl.clone();
-      url.pathname = "/login";
-      url.search = "";
-      return NextResponse.redirect(url);
-    }
-  }
-
-  return NextResponse.next();
-}
